003-n8n/paloalto-mcp-advanced/requirements.txt
fastmcp
pan-os-python
setuptools
